+++ packages/utils/package.json
@@ -1,6 +1,6 @@
 {
     "name": "@idearium/utils",
-    "version": "1.0.0-beta.2",
+    "version": "1.0.0-beta.3",
     "description": "Idearium utility methods.",
     "main": "index.js",
     "repository": "https://github.com/idearium/idearium-lib",
