+++ packages/utils/safe-promise.js
